refactor: simplify llm test suite
