--- front/lib/api/llm/clients/google/__test__/google.test.ts
@@ -1,75 +1,0 @@
-import { vi } from "vitest";
-
-import type {
-  GoogleAIStudioModelFamily,
-  GoogleAIStudioWhitelistedModelId,
-} from "@app/lib/api/llm/clients/google/types";
-import {
-  getGoogleAIStudioModelFamilyFromModelId,
-  GOOGLE_AI_STUDIO_WHITELISTED_MODEL_IDS,
-} from "@app/lib/api/llm/clients/google/types";
-import { TEST_CONVERSATIONS } from "@app/lib/api/llm/tests/conversations";
-import { LLMClientTestSuite } from "@app/lib/api/llm/tests/LLMClientTestSuite";
-import type {
-  ConfigParams,
-  TestConfig,
-  TestConversation,
-} from "@app/lib/api/llm/tests/types";
-import type { ModelIdType } from "@app/types/assistant/models/types";
-
-const GOOGLE_MODEL_FAMILY_TO_TEST_CONFIGS: Record<
-  GoogleAIStudioModelFamily,
-  ConfigParams[]
-> = {
-  "gemini-2": [
-    { reasoningEffort: "none" },
-    { reasoningEffort: "light", temperature: 0 },
-    { reasoningEffort: "medium", temperature: 1 },
-    { reasoningEffort: "high" },
-  ],
-  "gemini-3": [
-    { reasoningEffort: "none" },
-    { reasoningEffort: "light", temperature: 0 },
-    { reasoningEffort: "medium", temperature: 1 },
-    { reasoningEffort: "high" },
-  ],
-};
-
-// Inject Dust managed Google AI Studio credentials for testing
-vi.mock("@app/types", async (importOriginal) => {
-  const actual = await importOriginal();
-
-  return {
-    // @ts-expect-error actual is unknown
-    ...actual,
-    dustManagedCredentials: vi.fn(() => ({
-      GOOGLE_AI_STUDIO_API_KEY:
-        process.env.VITE_DUST_MANAGED_GOOGLE_AI_STUDIO_API_KEY ?? "",
-    })),
-  };
-});
-
-class GoogleTestSuite extends LLMClientTestSuite {
-  protected provider = "google_ai_studio" as const;
-  protected models = GOOGLE_AI_STUDIO_WHITELISTED_MODEL_IDS;
-
-  protected getTestConfig(
-    modelId: GoogleAIStudioWhitelistedModelId
-  ): TestConfig[] {
-    const family = getGoogleAIStudioModelFamilyFromModelId(modelId);
-    return GOOGLE_MODEL_FAMILY_TO_TEST_CONFIGS[family].map((configParams) => ({
-      ...configParams,
-      modelId,
-      provider: this.provider,
-    }));
-  }
-
-  protected getSupportedConversations(
-    _modelId: ModelIdType
-  ): TestConversation[] {
-    // Run all conversations for all Google models
-    return TEST_CONVERSATIONS;
-  }
-}
-
-new GoogleTestSuite().generateTests();
